.
ahora el importe total de cada venta se actualiza con con cada detalle de venta de mismo id.

+++ Venta.cpp
@@ -33,7 +33,7 @@ using namespace std;
 
     }
 
-    int Venta::SetImporteVenta(float v){
+    int Venta::SetImporteVenta(double v){
             if(v>0)
             {
                 _ImporteVenta=v;
